rpi / u-boot: disable TCP & wget
TCP implementation is too minimalistic and broken.
Receiving a TCP RST packet interrupts the running
transfer, even if the transfer command is not using
TCP (e.g., tftp)!
The TCP RST packet may be coming from an old connection
which was not explicitely closed last time the node
was UP. After a timeout (depending on TCP keepalive
parameters), the server detects the connection is dead,
so it closes it, which generates a TCP RST, and
sometimes interrupts the TFTP transfer if the node
is still booting.