config.txt: remove [pi1+] filter => fix pi2 boot!
The official doc now explicitely says that model B+ matches
the [pi1] filter; and it does not mention any [pi1+] filter.
With models now ordered from newest to oldest, the pi2 failed
to boot, probably because of the [pi1+] section which was
following the [pi2] section. The [pi1+] filter was probably
ignored (since it does not seem to be a valid filter) so the
board was considering it was still in section [pi2], and
it was overriding the kernel variable to "u-boot-rpi-b.img".

This commit removes the [pi1+] section. With this fix, pi2
boots properly again.